+++ tests/ByteDev.ResourceIdentifier.UnitTests/UriQueryConverterTests.cs
@@ -185,6 +185,17 @@ namespace ByteDev.ResourceIdentifier.UnitTests
                 Assert.That(result.GetValues("key1").Single(), Is.EqualTo("value1"));
                 Assert.That(result.GetValues("key2").Single(), Is.EqualTo("value2"));
             }
+
+            [Test]
+            public void WhenQueryNoValue_ThenReturnNameValueCollection()
+            {
+                var result = UriQueryConverter.ToNameValueCollection("?key1&key2&key3=value3");
+
+                Assert.That(result.AllKeys.Length, Is.EqualTo(3));
+                Assert.That(result.GetValues("key1"), Is.Null);
+                Assert.That(result.GetValues("key2"), Is.Null);
+                Assert.That(result.GetValues("key3").Single(), Is.EqualTo("value3"));
+            }
         }
 
         [TestFixture]
@@ -206,8 +217,8 @@ namespace ByteDev.ResourceIdentifier.UnitTests
                 var result = UriQueryConverter.ToDictionary("?key1=value1&key2=value2");
 
                 Assert.That(result.Count, Is.EqualTo(2));
-                Assert.That(result.First().Value, Is.EqualTo("value1"));
-                Assert.That(result.Second().Value, Is.EqualTo("value2"));
+                Assert.That(result["key1"], Is.EqualTo("value1"));
+                Assert.That(result["key2"], Is.EqualTo("value2"));
             }
 
             [Test]
@@ -216,8 +227,19 @@ namespace ByteDev.ResourceIdentifier.UnitTests
                 var result = UriQueryConverter.ToDictionary("key1=value1&key2=value2");
 
                 Assert.That(result.Count, Is.EqualTo(2));
-                Assert.That(result.First().Value, Is.EqualTo("value1"));
-                Assert.That(result.Second().Value, Is.EqualTo("value2"));
+                Assert.That(result["key1"], Is.EqualTo("value1"));
+                Assert.That(result["key2"], Is.EqualTo("value2"));
+            }
+
+            [Test]
+            public void WhenQueryNoValue_ThenReturnNameValueCollection()
+            {
+                var result = UriQueryConverter.ToDictionary("?key1&key2&key3=value3");
+
+                Assert.That(result.Count, Is.EqualTo(3));
+                Assert.That(result["key1"], Is.Null);
+                Assert.That(result["key2"], Is.Null);
+                Assert.That(result["key3"], Is.EqualTo("value3"));
             }
         }
     }