+++ docs/RELEASE-NOTES.md
@@ -1,12 +1,24 @@
 # Release Notes
 
+## 1.2.0 - ?
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `UriSetExtensions.SetQuery` overload for `IEnumerable<string>`.
+
+Bug fixes / internal changes:
+- Fixed `UriQueryConverter.ToNameValueCollection` implementation to handle query string name params with no value.
+- Fixed `UriQueryConverter.ToDictionary` implementation to handle query string name params with no value.
+
 ## 1.1.0 - 11 December 2020
 
 Breaking changes:
 - (None)
 
 New features:
-- Added overload for `UriQueryConverter.ToString`
+- Added overload for `UriQueryConverter.ToString`.
 
 Bug fixes / internal changes:
 - (None)